requirements.txt
langchain-groq
streamlit
python-dotenv
langchain-huggingface